tree maps with color scale

+++ Sidi.Util/Visualization/LabelPainter.cs
@@ -139,12 +139,12 @@ namespace Sidi.Visualization
 
         bool focusPointEnabled = false;
         Point focusPoint;
-        float[] worldFocusPoint;
+        double[] worldFocusPoint;
 
         public void Paint(PaintEventArgs e)
         {
             var layout = treeMapControl.LayoutManager;
-            float maxArea = layout.Root.Bounds.Area();
+            double maxArea = layout.Root.Bounds.Area();
             alphaF = 220.0 / (Math.Log10(maxArea) - Math.Log10(MinArea));
             if (focusPointEnabled)
             {
